+++ assets/vue/components/settings/SettingsActionsPanel.vue
@@ -2,15 +2,9 @@
   <section class="space-y-6">
     <div class="flex items-center justify-between">
       <div>
-        <h2 class="text-xl font-bold text-slate-900">Bounce Management</h2>
-        <p class="mt-0.5 text-sm text-slate-500">
-          Monitor, process and automate bounce handling across all campaigns
-        </p>
+        <h2 class="text-xl font-bold text-slate-900">Settings</h2>
+        <p class="mt-0.5 text-sm text-slate-500">Application configuration and admin management</p>
       </div>
-<!--      <div class="hidden sm:flex items-center gap-1.5 px-3 py-1.5 bg-emerald-50 border border-emerald-200 rounded-lg">-->
-<!--        <span class="w-2 h-2 rounded-full bg-emerald-500" />-->
-<!--        <span class="text-xs font-medium text-emerald-700">0.8% Bounce Rate</span>-->
-<!--      </div>-->
     </div>
 
     <div class="flex flex-wrap gap-2">
@@ -25,32 +19,25 @@
         @click="setActiveTab(tab.id)"
       >
         <span>{{ tab.label }}</span>
-        <span
-          v-if="tab.count"
-          class="ml-2 text-[10px] px-1.5 py-0.5 rounded-full font-semibold"
-          :class="activeTab === tab.id ? 'bg-blue-100 text-blue-700' : 'bg-slate-100 text-slate-500'"
-        >
-          {{ tab.count }}
-        </span>
       </button>
     </div>
 
-    <template v-if="activeTab === 'overview'">
-      <BounceOverview />
+    <template v-if="activeTab === 'configs'">
+      <SettingsConfigs />
     </template>
 
-    <template v-else-if="activeTab === 'rules'">
-      <BounceRules />
+    <template v-else-if="activeTab === 'admins'">
+      <SettingsAdmins />
     </template>
 
-    <template v-else-if="activeTab === 'per'">
-      <BouncePer />
+    <template v-else-if="activeTab === 'admin_attributes'">
+      <SettingsAdminAttributes />
     </template>
 
     <template v-else>
       <div class="bg-white rounded-xl border border-slate-200 p-6">
         <h3 class="text-base font-semibold text-slate-900">{{ currentTabLabel }}</h3>
-        <p class="mt-1 text-sm text-slate-500">This panel matches the template navigation and can be expanded next.</p>
+        <p class="mt-1 text-sm text-slate-500">This panel is under construction.</p>
       </div>
     </template>
   </section>
@@ -59,23 +46,20 @@
 <script setup>
 import { computed, ref, watch } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
-import BounceOverview from './BounceOverview.vue'
-import BounceRules from './BounceRules.vue'
-import BouncePer from "./BouncePer.vue";
+import SettingsConfigs from './SettingsConfigs.vue'
+import SettingsAdmins from './SettingsAdmins.vue'
+import SettingsAdminAttributes from './SettingsAdminAttributes.vue'
 
 const tabs = [
-  { id: 'overview', label: 'Overview' },
-  { id: 'rules', label: 'Rules' },
-  { id: 'per', label: 'Per List/Campaign' },
-  { id: 'processing', label: 'Processing' },
-  // { id: 'per_list', label: 'Per List' },
-  // { id: 'per_campaign', label: 'Per Campaign' },
+  { id: 'configs', label: 'Configs' },
+  { id: 'admins', label: 'Admins' },
+  { id: 'admin_attributes', label: 'Admin Attributes' },
 ]
 
 const route = useRoute()
 const router = useRouter()
 
-const defaultTabId = 'overview'
+const defaultTabId = 'configs'
 const tabIds = new Set(tabs.map((tab) => tab.id))
 
 const normalizeTabId = (tabId) => (typeof tabId === 'string' && tabIds.has(tabId) ? tabId : defaultTabId)
@@ -115,3 +99,4 @@ watch(
 
 const currentTabLabel = computed(() => tabs.find((tab) => tab.id === activeTab.value)?.label || 'Panel')
 </script>
+
