+++ assets/vue/components/settings/SettingsAdminAttributes.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="bg-white rounded-xl border border-slate-200 p-6">
+    <h3 class="text-base font-semibold text-slate-900">Admin Attributes</h3>
+    <p class="mt-1 text-sm text-slate-500">Custom attributes for admin profiles.</p>
+
+    <!-- Placeholder: implement attributes list and editing -->
+    <div class="mt-4 text-sm text-slate-500">Admin attributes UI coming soon.</div>
+  </div>
+</template>
+
+<script setup>
+// Future: implement admin attribute CRUD
+</script>
+
