+++ assets/vue/components/settings/SettingsAdmins.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="bg-white rounded-xl border border-slate-200 p-6">
+    <h3 class="text-base font-semibold text-slate-900">Admins</h3>
+    <p class="mt-1 text-sm text-slate-500">Manage administrator accounts and permissions.</p>
+
+    <!-- Placeholder: implement admins list, create/edit actions -->
+    <div class="mt-4 text-sm text-slate-500">Admins management UI coming soon.</div>
+  </div>
+</template>
+
+<script setup>
+// Future: implement admin list, create/edit/delete actions
+</script>
+
